00-json-transformation-service-main/json-transformation-service-main/json-transformation-core/src/main/java/com/navi/bootcamp/json/transformation/core/utils/CommonUtils.java
package com.navi.bootcamp.json.transformation.core.utils;

import com.fasterxml.jackson.core.JsonProcessingException;
import com.fasterxml.jackson.databind.ObjectMapper;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Component;

@Component
@RequiredArgsConstructor
public class CommonUtils {

    private final ObjectMapper objectMapper;

    public String objectToString(Object o) {
        try {
            return objectMapper.writeValueAsString(o);
        } catch (JsonProcessingException e) {
            throw new RuntimeException("object to string conversion failed", e);
        }
    }
}
